docker/requirement.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.15.0
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2024.1
